+++ docs/realm.js
@@ -455,9 +455,10 @@ class Realm {
  *   This is not supported for `"list"` properties of object types and `"linkingObjects"` properties.
  * @property {boolean} [indexed] - Signals if this property should be indexed. Only supported for
  *   `"string"`, `"int"`, and `"bool"` properties.
- * @property {string} [alias] - an alias in JavaScript for the underlying name in the Realm file. If set, this must be
- *   used when reading/writing the property or making queries against it. Trying to use the internal name will throw
- *   an exception.
+ * @property {string} [mapTo] - Set this if the Javascript property name schema differs from the underlying name in the Realm
+ *   file. This can be used to have different naming convention in Javascript than what is being used in the file
+ *   itself. Reading and writing properties must be done using the public key for the property. Queries can be
+ *   done using both the public and the mapped property name.
  */
 
 /**