+++ lib/index.d.ts
@@ -49,7 +49,7 @@ declare namespace Realm {
         default?: any;
         optional?: boolean;
         indexed?: boolean;
-        alias?: string;
+        mapTo?: string;
     }
 
     // properties types